change day name to short

+++ weather-app/src/helpers/get-day-name.js
@@ -1,6 +1,6 @@
 const getDayName = (time, locale) => {
     let date = new Date(time);
-    return date.toLocaleDateString(locale, { weekday: 'long' });
+    return date.toLocaleDateString(locale, { weekday: 'short' });
 };
 
 export default getDayName;
